requirements.txt
requests>=2.25.0
websocket-client>=1.2.0
pandas>=1.3.0
numpy>=1.20.0
pytz>=2021.1
schedule>=1.1.0
pyOpenSSL>=25.0.0
twisted>=25.5.0

# Trading APIs
kiteconnect>=5.0.0

# Trading and Technical Analysis
scikit-learn>=1.0.0
ta-lib>=0.4.0
yfinance>=0.1.70

# CLI and User Interface
click>=8.0.0
tabulate>=0.9.0
colorama>=0.4.6

# Database Support
# ClickHouse (Recommended for time series)
clickhouse-driver>=0.2.6
clickhouse-connect>=0.6.0

# PostgreSQL (Alternative)
psycopg2-binary>=2.9.0
SQLAlchemy>=1.4.0
alembic>=1.8.0

# Redis for caching
redis>=4.5.0
hiredis>=2.0.0

# News Agent Dependencies
aiohttp>=3.8.0
feedparser>=6.0.0

# Development and Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
matplotlib>=3.5.0
# mplfinance releases on PyPI are currently pre-releases; install the
# latest stable code directly from the GitHub repository. If you prefer
# the PyPI pre-release, run `pip install --pre mplfinance` instead.
mplfinance @ git+https://github.com/matplotlib/mplfinance.git@master